+++ src/test/java/org/apiphany/security/oauth2/OAuth2TokenProviderOptionsTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.morphix.reflection.Constructors;
 
 /**
- * Test class for {@link OAuth2TokenProviderOptions}.
+ * Test class for {@link OAuth2TokenProviderProperties}.
  *
  * @author Radu Sebastian LAZIN
  */
@@ -17,7 +17,7 @@ class OAuth2TokenProviderOptionsTest {
 	@Test
 	void shouldNotInstantiateDefaultClass() {
 		UnsupportedOperationException unsupportedOperationException =
-				Tests.verifyDefaultConstructorThrows(OAuth2TokenProviderOptions.Default.class);
+				Tests.verifyDefaultConstructorThrows(OAuth2TokenProviderProperties.Default.class);
 		assertThat(unsupportedOperationException.getMessage(), equalTo(Constructors.MESSAGE_THIS_CLASS_SHOULD_NOT_BE_INSTANTIATED));
 	}