+++ src/test/java/org/apiphany/utils/client/DummyExchangeClient.java
@@ -4,6 +4,13 @@ import org.apiphany.ApiRequest;
 import org.apiphany.ApiResponse;
 import org.apiphany.client.ExchangeClient;
 
+/**
+ * Dummy implementation of {@link ExchangeClient} for testing purposes.
+ * <p>
+ * Only the abstract methods are implemented with default behavior.
+ *
+ * @author Radu Sebastian LAZIN
+ */
 public class DummyExchangeClient implements ExchangeClient {
 
 	@Override